+++ recruiter_webapp/Skill.aspx.cs
@@ -26,20 +26,30 @@ namespace recruiter_webapp
 
         protected void Page_Load(object sender, EventArgs e)
         {
-            lblResponseMsg.Text = "";
-            if (!IsPostBack)
+            if (Session["user_id"] != null)
             {
-                ApiPath = ConfigurationManager.AppSettings["Api"].ToString();
-                WebURL = ConfigurationManager.AppSettings["WebURL"].ToString();
+                if (Convert.ToInt32(Session["user_id"]) < 6)
+                {
+                    lblResponseMsg.Text = "";
+                    if (!IsPostBack)
+                    {
+                        ApiPath = ConfigurationManager.AppSettings["Api"].ToString();
+                        WebURL = ConfigurationManager.AppSettings["WebURL"].ToString();
+                    }
+
+                    if (Request.Url.ToString().Contains("Delete"))
+                        DeleteSkill();
+
+                    /* if (Request.Url.ToString().Contains("Edit"))
+                         EditSkill(); */
+
+                    GetSkills();
+                }
             }
-
-            if (Request.Url.ToString().Contains("Delete"))
-                DeleteSkill();
-
-           /* if (Request.Url.ToString().Contains("Edit"))
-                EditSkill(); */
-
-            GetSkills();
+            else
+            {
+                Response.Redirect(ConfigurationManager.AppSettings["WebURL"].ToString());
+            }
         }
